Update README.md
New feature:  ClickScan can now be activated from an iPad switch interface. Set the interface like the Pretorian iSwitch 75 to issue ~1 when the user presses the switch. By using a Bluetooth switch interface you gain the advantages of wireless access, and you do not need to hover the mouse cursor over the red button on the ClickScan window.